rename: change package name
Change package name from bridgeadd to bridgemessageadd.
