+++ src/main/java/bridge/domain/bridgemessageadd/UpBridgeMessageAddStrategy.java
@@ -1,4 +1,4 @@
-package bridge.domain.bridgeadd;
+package bridge.domain.bridgemessageadd;
 
 class UpBridgeMessageAddStrategy implements BridgeMessageAddStrategy {
     public void addMessage(String message, BridgeMessageAdder bridgeMessageAdder) {
